ajustement
just a small bugs fixed
