+++ src/page/KeyLogger.js
@@ -28,7 +28,7 @@ export default function KeyLogger(props) {
     }, [phone])
 
     return <div className="card p-3"  style={{ height: 'auto'}}>
-        <h1>Log d'appels</h1>
+        <h1>Saisie du clavier</h1>
 
         <table className="table">
             <thead>
