+++ src/page/style.js
@@ -1,0 +1,52 @@
+const customCourbes = {
+    display: "flex",
+    justifyContent: "space-between",
+    padding: "20px"
+}
+
+const customCourbesDetails = {
+    display: "flex",
+    justifyContent: "space-between",
+    alignItems: "center"
+}
+
+const customCourbesTitle = {
+    fontSize: "20px",
+    fontWeight: "500",
+    textAlign: "left"
+}
+
+const customStyle = {
+    width: "65%",
+    background: "#fff",
+    borderRadius: "12px",
+    boxShadow: "0 5px 10px rgba(0,0,0,0.1)",
+    padding: "20px",
+    
+}
+
+const customStats = {
+    width: "35%",
+    marginLeft: "20px",
+    background: "#fff",
+    boxShadow: "0 5px 10px rgba(0,0,0,0.1)",
+    padding: "20px",
+    borderRadius: "12px"
+}
+
+const customGps = {
+    width: "auto",
+    marginLeft: "20px",
+    background: "#fff",
+    boxShadow: "0 5px 10px rgba(0,0,0,0.1)",
+    padding: "20px",
+    borderRadius: "12px"
+}
+
+const customGpsTitle = {
+    fontSize: "20px",
+    fontWeight: "500",
+    textAlign: "center"
+}
+
+export {customCourbes, customCourbesDetails, customCourbesTitle, customStyle, customStats, customGps};
